Centralize cache health checks and add timeouts
Remove duplicate CacheHealthCheck implementations from Account and Admin.BFF and add a shared CacheHealthCheck (DotNet/Shared/Application/Health) that uses a 5s HealthCheckTimeout and includes exception details on failure. Update ResourceCacheHealthCheck to use the same 5s timeout and to pass cancellation tokens to Redis Ping and Blob storage checks. Also update Admin.BFF MeasureEvalService and ValidationService to apply a 5s HealthCheckTimeout by creating linked CancellationTokenSources and canceling HTTP health GETs after the timeout. These changes centralize health-check logic and prevent long-running/hanging cache and HTTP calls.
