+++ DotNet/Admin.BFF/Application/Clients/ValidationService.cs
@@ -15,6 +15,7 @@ public class ValidationService
     private readonly HttpClient _client;
     private readonly IOptions<ServiceRegistry> _serviceRegistry;
     private const string HealthUp = "UP";
+    private static readonly TimeSpan HealthCheckTimeout = TimeSpan.FromSeconds(5);
 
 
     public ValidationService(ILogger<ValidationService> logger, HttpClient client, IOptions<ServiceRegistry> serviceRegistry)
@@ -42,6 +43,9 @@ public class ValidationService
 
     public async Task<LinkServiceHealthReport> LinkServiceHealthCheck(CancellationToken cancellationToken)
     {
+        using var timeoutCts = CancellationTokenSource.CreateLinkedTokenSource(cancellationToken);
+        timeoutCts.CancelAfter(HealthCheckTimeout);
+
         // HTTP GET
 
         var report = new LinkServiceHealthReport
@@ -51,7 +55,7 @@ public class ValidationService
 
         try
         {
-            var response = await _client.GetAsync($"health", cancellationToken);
+            var response = await _client.GetAsync($"health", timeoutCts.Token);
 
             var content = await response.Content.ReadAsStringAsync();
